+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/diffbased/common/storage/flat/CodeHashCodeStorageStrategy.java
@@ -34,6 +34,7 @@ public class CodeHashCodeStorageStrategy implements CodeStorageStrategy {
 
   @Override
   public void putFlatCode(
+      final SegmentedKeyValueStorage storage,
       final SegmentedKeyValueStorageTransaction transaction,
       final Hash accountHash,
       final Hash codeHash,
@@ -43,6 +44,7 @@ public class CodeHashCodeStorageStrategy implements CodeStorageStrategy {
 
   @Override
   public void removeFlatCode(
+      final SegmentedKeyValueStorage storage,
       final SegmentedKeyValueStorageTransaction transaction,
       final Hash accountHash,
       final Hash codeHash) {}