+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/diffbased/common/storage/flat/FlatDbStrategyProvider.java
@@ -21,6 +21,7 @@ import static org.hyperledger.besu.ethereum.trie.diffbased.common.storage.DiffBa
 import org.hyperledger.besu.ethereum.worldstate.DataStorageConfiguration;
 import org.hyperledger.besu.ethereum.worldstate.FlatDbMode;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
+import org.hyperledger.besu.plugin.services.storage.DataStorageFormat;
 import org.hyperledger.besu.plugin.services.storage.SegmentedKeyValueStorage;
 
 import java.nio.charset.StandardCharsets;
@@ -36,8 +37,8 @@ public abstract class FlatDbStrategyProvider {
 
   // 0x666C61744462537461747573
   public static final byte[] FLAT_DB_MODE = "flatDbStatus".getBytes(StandardCharsets.UTF_8);
-  private final MetricsSystem metricsSystem;
-  private final DataStorageConfiguration dataStorageConfiguration;
+  protected final MetricsSystem metricsSystem;
+  protected final DataStorageConfiguration dataStorageConfiguration;
   protected FlatDbMode flatDbMode;
   protected FlatDbStrategy flatDbStrategy;
 
@@ -99,7 +100,17 @@ public abstract class FlatDbStrategyProvider {
   @VisibleForTesting
   private FlatDbMode deriveFlatDbStrategy(
       final SegmentedKeyValueStorage composedWorldStateStorage) {
-    final FlatDbMode requestedFlatDbMode = getRequestedFlatDbMode(dataStorageConfiguration);
+    final FlatDbMode requestedFlatDbMode =
+        dataStorageConfiguration
+                .getDiffBasedSubStorageConfiguration()
+                .getUnstable()
+                .getFullFlatDbEnabled()
+            ? (dataStorageConfiguration
+                    .getDataStorageFormat()
+                    .equals(DataStorageFormat.X_BONSAI_ARCHIVE)
+                ? FlatDbMode.ARCHIVE
+                : FlatDbMode.FULL)
+            : FlatDbMode.PARTIAL;
 
     final var existingTrieData =
         composedWorldStateStorage.get(TRIE_BRANCH_STORAGE, WORLD_ROOT_HASH_KEY).isPresent();
@@ -112,8 +123,8 @@ public abstract class FlatDbStrategyProvider {
                 .orElseGet(
                     () -> {
                       // if we do not have a db-supplied config for flatdb, derive it:
-                      // default to partial if trie data exists, but the flat config does not,
-                      // and default to the storage config otherwise
+                      //  - default to partial if trie data exists but the flat config does not,
+                      //  - otherwise go with the requested mode
                       var flatDbModeVal =
                           existingTrieData
                               ? alternativeFlatDbModeForExistingDatabase().getVersion()
@@ -142,9 +153,6 @@ public abstract class FlatDbStrategyProvider {
   public FlatDbMode getFlatDbMode() {
     return flatDbMode;
   }
-
-  protected abstract FlatDbMode getRequestedFlatDbMode(
-      final DataStorageConfiguration dataStorageConfiguration);
 
   protected abstract FlatDbMode alternativeFlatDbModeForExistingDatabase();